added authentication to web app

+++ src/components/TaskDetails.js
@@ -1,10 +1,22 @@
 import { useTasksContext } from "../hooks/useTasksContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const TaskDetails = ({task}) => {
+
     const { dispatch } = useTasksContext()
+    const { user } = useAuthContext()
+
     const handleClick = async () => {
+
+        if (!user){
+            return
+        }
+
         const response = await fetch('https://apicrud-n1uz.onrender.com/api/tasks/' + task._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
